# feat(progress bar): Gives constant execution feedback