--- a/main.py
+++ b/main.py
@@ -3,3 +3,7 @@ from src.cli.commands import execute
 
 def run():
     execute()
+
+
+if __name__ == "__main__":
+    run()