--- a/setup.py
+++ b/setup.py
@@ -45,6 +45,7 @@ setup(
         "faker>=4.0.1",
         "sqlalchemy>=1.3.16",
         "mysql-connector-python>=8.0.19",
+        "alive-progress>=1.5.1",
         "azure-storage-blob>=12.3.0",
         "psycopg2-binary>=2.8.5"
     ],