@macunha1 2020-06-07
Performance benchmark results:

Without progress bar
```bash
> time ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json
ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json  22.54s user 0.46s system 101% cpu 22.761 total

> time ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json
ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json  22.10s user 0.46s system 101% cpu 22.256 total

> time ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json
ggenerator generate --spec /home/macunha1/Downloads/ggenerator/spec.json  22.11s user 0.55s system 101% cpu 22.364 total
```

With it
```bash
> time python /home/macunha1/Workspace/Datenworks/Public/ggenerator/main.py generate --spec spec.json
python /home/macunha1/Workspace/Datenworks/Public/ggenerator/main.py generate  22.94s user 0.56s system 101% cpu 23.241 total

> time python /home/macunha1/Workspace/Datenworks/Public/ggenerator/main.py generate --spec spec.json
python /home/macunha1/Workspace/Datenworks/Public/ggenerator/main.py generate  23.29s user 0.51s system 100% cpu 23.568 total
```